Brown paper bag bug: setting HOST_BUILD_EXTRA or using ./download.sh --extract causes a build break.
The regex in the shell function noversion doesn't recognize two consecutive letters as part of a legitimate version name, so qemu-2d18e637e5ec.tar.bz2 should actually be qemu-2d18e637e5e.tar.bz2 in order to be recognized as "qemu".  (That's using the git commit ID as the version, which would have worked if I'd cut it in the right place.)

I forgot that ./buildall.sh doesn't use ./download.sh --extract anymore, so of course it didn't catch this.