Fix source/target confusion that bites Fedora host = target builds.
The gcc build exports an LD_LIBRARY_PATH full of target libraries and then
calls /bin/sh in the configure stage. If you configure before building, this
doesn't cause a problem because the libraries aren't there yet, but c++ has
dependencies that trigger a C build before running configuration even if you
reorder the steps. On ubuntu this doesn't cause a problem because /bin/sh
doesn't link against anything in this LD_LIBRARY PATH, and when building
a different architecture the foreign binaries are ignored by the dynamic linker.
But on Fedora when host == target /bin/sh links against libgcc_s.so.1 which is
in LD_LIBRARY_PATH linked against libc.so.0 which isn't installed on the
host, and the attempt to run the shell dies saying it can't find uClibc.

The fix is to override the make variable SHELL = sh so it calls ash out of
build/host.